Allow selecting the image resolution instead of "small/midium/large"
The concept of small, midium, large is relative, as it depends on
whatever the camera supports.

Instead of using it, it is a way better to show a list of the
supported resolutions. That works fine for most cameras, with
uses a fixed set.

On cameras with a scale, the Kernel actually exports the resolution
range. On this case, let's present 4 resolutions between the
minimum and maximum one. That should be reasonable for most
usecases.

Nowadays, all drivers should be reporting resolutions via
VIDIOC_ENUM_FRAMESIZES. Yet, as one might run camorama with
very legacy kernels, provide backward support, using
small/midium/large resolutions just like before.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>