v4l: make print_cam() useful
There is a debug logic at v4l.c that is not used anywhere.
Also, its output is noisy, as it displays some things that
shouldn't work if not explicitly set.

Print data from it when -D is used, on a useful way, e. g.
by default it will print something like:

    Camera Info
    -----------
    device = /dev/video0, x = 960, y = 720
    bits per pixel = 24, desk_depth = 24
    capture directory = /home/user/Webcam_Pictures, capture file = Webcam

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>