camorama.glade: use glade-3 to generate it
It is a way easier to use glade-3 to change the interface.
Now that it doesn't use the obsoleted stock item anymore,
convert it to the format that glade-3 uses.

It will produce one warning, though:
	[main_window:bonobodock1:vbox36:hbox25:frame1:da] Object class 'Custom widget' from gtk 2.16 is deprecated

It should be noticed that glade-3 will add a "swapped attribute
to all signals. Those will cause harmless warnings when
camorama runs. In order to avoid, we just need to patch the
file, after using glade-3 with hunks like this:

-                            <signal name="activate" handler="on_show_effects_activate" swapped="no"/>
+                            <signal name="activate" handler="on_show_effects_activate"/>

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>