Update camorama.desktop.in
That basically adds a change that went to Fedora Camorama
packaging. I was in doubt about that, so I went to Freedesktop
in order to check the current standard:

	https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html

I ordered the values as placed there, in order to make easier to
validate it with the spec.

The Category now follows desktop the standard, as specified at:
	https://standards.freedesktop.org/menu-spec/latest/apa.html

As Video category should include AudioVideo, and apps based on
Gnome libraries should have GNOME.

From the example:
	https://standards.freedesktop.org/menu-spec/latest/apcs03.html

Icons don't have extension anymore. Fedora packaging complains about
that as well. So, the .png from camorama icon got removed.

Also added a Keywords.

Finally, as we're not using gnome BZ anymore, remove the entries
related to them.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>