Unify coding style
The coding style of Camorama is a mess. It doesn't consistently
follow the same style. Unify it by applying a coding style
close to the Linux Kernel. There is an exception, however:
here, we'll use indent with 4, and won't use tabs.

Rationale for using the Kernel style:
1) The checkpatch.pl script there can do lots of cleanups;
2) There used to be a lindent line (with I still have here),
   with do several other style cleanups.

So, most of the changes can happen on a more automatic way.
Still, care should be taken with static constants, as Kernel
initializes them with zeros, while userspace don't.

Accidentally, I'm more familiar with this style, with makes
easier for me to contribute :-D

No functional changes.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>