add support for scaling the image with cairo
It is easy to scale the image with cairo. Add support for it.

For now, it reserves about 2/3 of the screen for the image stream
display, keeping 1/3 reserved for the interface.

It probably makes sense to add ways for the user to change the
scale.

It should be noticed that the snapshots won't be affected for
the scale; will keep using the image resolution.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>