callbacks: use a callback to draw at GtkDrawingArea
Gtk 3 supports wayland, but there's a catch: while the
current code works properly with X11, it doesn't draw
anything inside the GtkDrawingArea on Wayland.

The fix is trivial, but incompatible with Gtk2: instead of
updating the drawing aea inside the main loop, a callback
has to be used in order to handle gtk draw logic.

As, at least for now, we want to keep supporting Gtk 2,
let's add a compatibility code.

That fixes drawing with Wayland.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>