callbacks: fix apply_filters() logic
The filters should work with RGB data. If data is on YUV420,
they should be converted first to RGB before filtered.

Please notice that this shouldn't happen, in practice, as
libv4l will always provide a RGB output, but, if something
changes there, it may end doing a wrong thing. So, better
to fix.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>