Remove the tray icon
There is a somewhat complex logic here that just adds a
camorama icon to the tray bar. The only action it provides is
that a click makes the window to minimize.

That's too much effort for too little gain. As the API it
uses is deprecated on gtk 3 in favor of Gnotification, it
sounds better to just get rid of it for good.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>