add the compatibility bits needed to build it with gtk 4
Now that we don't use gtk_init_*() anymore, there are just a few
minor changes required at the source code in order to make it
build with Gtk 4.

Please notice that there's a small hack there: originally, the
popup menu required mouse button 3. I was unable to discover
how to get it with gtk 4. As the other buttons are unused, with
the new code, any mouse button will open the popup effects
window.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>