main: make it use gtk_application_new() & friends
As with gtk 4 there's no way to pass command line arguments anymore
to gtk_init, we need to change the logic, in order to use
g_application() instead.

It should be noticed that g_application() is not supported
by gtk 2. So, this patch will temporarely break compatibility
with it. The next patch will restore.

I opted to add the backward compat on a separate patch as it
makes easier to later drop gtk 2 compat, after we test gtk 3
support for a while.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>