Get rid of obsolete Gtk3 widgets
On Gtk3, several Widgets got obsoleted, being replaced by
a different wigdet.

The only exception is the GtkImageMenu item, that were simply
removed, without any replacement. It seems that now it is not
possible anymore (or at least it is not trivial) to associate
a MenuItem with an image.

Anyway, as those widgets don't exist anymore on Gtk4, we'll
need to replace them anyway.

The new camorama-gtk3.ui now validates with Glade 3.22.
It should not work with gtk2, so, at least while we keep
backport compatibility, let's have both schemas.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>