camorama-gtk3: explicitly use the obsoleted GtkImageMenuItem
The Gtk 3 manual says it would be possible to use a GtkBox
with a GtkLabel and GtkImage in order to replace it.

I tried, both directly and using Glade 3.22. Didn't work.

So, let's explicitly revert this change, as it makes
Camorama with a bad visual (IMHO).

We'll still need some solution if we want to preserve icons
on menus for Gtk 4, but that's a separate issue.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>