interface.h: rename Gconf2 keys
Using KEY<number> makes harder to identify the meaning of it,
as KEY* macros are also used at Linux input/evdev subsystem.

Also, when something changes there, we start by having numbers
skipped or needing to place stuff out of order.

Finally, "PATH" is a terrible name, as it usually means the
environment var.

So, rename everything inside interface.h, in order to make
easier to use them, and clearer that they're related to
Gconf2.

While here, get rid of "date_format", as it is not used
anywhere.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>