Video devices: improve the detection logic
The previous detection logic were unable to exclude devices
that weren't provide video capture streams. It were also
checking if the device is really a V4L based at the device
name.

The new logic now uses VIDIOC_QUERYCAP to check if the device
is really a V4L2 and checks if the specific device node is
capable of doing video capture.

There's one big advantage with this strategy: if udev
decides to not follow the usual convention, this will still
work.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>