add support for full screen mode
Nowadays, there are several UVC devices that can be used to
capture HDMI images. Those could be used, for instance, to
display a console from another machine. On such case, it can
be interesting to use as little as possible of the screen,
in order to maximize the screen.

Let's implement a full screen mode that will disable everything
but the top menu, and place the application in full screen
mode.

As this depends on a proper scaling support, only enable it
with Gtk3.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>