Add support for userptr
The V4L2 API supports 3 streaming methods, but only two of them
are supported currently by camorama. With the userptr method,
userspace allocates memory, instead of the Kernel.

This is supported by most drivers, but the staging atomisp one
currently doesn't work with MMAP. This should be fixed on some
later version, but adding userptr support on camorama would
allow testing other parts of the driver.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>