v4l: use depth when selecting the video format
On cameras, usually the USB bus limits the bandwidth, which affects
the maximum number of frames per second. So, formats with smaller
depth are usually faster.

When the cameras report the max speed via ENUM_FRAMEINTERVALS,
this is already taken in consideration. However, not all cameras
report it. So, take the depth of the video format into account.

This would make camorama to select the format with the smallest
depth, among the ones it recognizes, hopefully allowing a better
speed.

It should also be noticed that modern cameras use compressed
formats like MJPEG in order to overcome USB2 bus limitation.

As MJPEG is not listed at the formats table, it would mean
depth = 0, which would make camorama to select it by default,
if libv4l2 is used or if we end adding support for it natively
in the future.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>