Avoid a tmp buffer when not using read()
When mmap or userptr is used, there's no need for an extra
buffer, as:
	- the DQBUF has the raw data;
	- cam->pix_buf has size for RGB24.

That's enough, as, if libv4l2 is used, the buffer is already
in RGB24. So, a simple copy (dropping pads) is needed.

When libv4l2 is not used, the logic just needs to fill the
data directly into the pix_buf from que buffer queue.

That avoids extra buffering, speeding up the speed when
libv4l2 is not used.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>