TODO: remove items that were already done
While we're not using the TODO list there anymore, it is outdated,
as some items were already addressed. Remove them from the TODO
list.

There is one pending item that I may end implementing some day:
support for video recording. It won't be record it locally saving
the video in MJPEG, for example, but, except for video monitoring,
in order for such feature to be useful, Camorama would need to also
record audio, which is a little trickier.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>