fileio: allow storing more than one snapshot per second
Currently, as the image name carries the date in seconds, it is
not possible to have more than one snapshot per second, as otherwise
the second one would override the initial one.

There are several ways to solve it, but, as the minimum granularity
is the frame rate, use a frame counter, writing the least 3 digits
of it at the file name. That would allow it to scale up to 1000fps,
which is a lot more than the normal webcams would allow.

Add another number at the saved file name, in order to

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>