drop gtk2 and legacy gtk3 support
Gtk 3.22 was released in 2016, which is old enough to be even
on LTS distros. Also, the current code was not tested - and doesn't
actually build with Gtk2.

While it would be doable to make it work again with Gtk2, what's
the point? Let's just move forward and setup the minimal to be
Gtk 3.22.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>