README.md, NEWS: update information about Threshold
The Threshold filter classes (both overall and per channel)
have sets "self->threshold = 127". Yet, camorama README.md
says that the threshold could be adjusted via a slidebar.

This is probably part of some future implementation, or the
original implementation got lost in time.

Either way, the current code fixes the limiar to 127.

In order to be able to change that, it would be required to
change the filter class init code to create or unhide new
slide bars, and add a finalize code that would hide or delete
the slide if the filter is removed.

Update it accordingly at the README.md, and add a note at the
TODO about implementing support for per-filter slidebar(s).

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>