callbacks: reflect changes from control window at main window
If one changes a control at the "Show Controls" window, reflect it
also at the main window.

Ideally, we should do the reverse too, but this would be complex.
Not sure if it would be worth. Also, if we implement events support
for the "Show Controls" window, this would come for free.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>