Fix crashes when used with atomisp2 cameras
Buffers allocated for USERPTR buffers must always be PAGE_SIZE
aligned and their size must always be round up to a multiple
of PAGE_SIZE.

This fixes various camorama crashes when used with atomisp2 cameras,
likely caused by the atomisp DMA-ing over malloc's internal data
structures which usually are stored before the actual allocated
memory.

And as an added bonus it also fixes what appeared to be hsync issues
at certain resolutions, but in reality was calloc returning a not
PAGE_SIZE aligned pointer, where as the atomisp2 was writing data
starting at the start of the page.

With this + some kernel fixes I can switch resolutions on an ov2680
without any visual issues or crashes.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>